+++ src/api/clipboard.ts
@@ -2,7 +2,7 @@ import { StoreType, AreaType } from "../types";
 
 import { zoneToArea } from "./structs";
 import { matrix2tsv } from "./converters";
-import { solveTable } from "../formula/evaluator";
+import { solveTable } from "../formula/solver";
 
 export const clip = (store: StoreType): AreaType => {
   const { selectingZone, choosing, editorRef, table } = store;
