+++ src/formula/functions/countif.ts
@@ -1,4 +1,5 @@
-import { solveTable, FormulaError } from "../evaluator";
+import { FormulaError } from "../evaluator";
+import { solveTable } from "../solver";
 import { Table } from "../../api/table";
 import { BaseFunction } from "./__base";
 import { check } from "./__utils";
