+++ src/formula/functions/vlookup.ts
@@ -1,4 +1,5 @@
-import { solveTable, FormulaError } from "../evaluator";
+import { FormulaError } from "../evaluator";
+import { solveTable } from "../solver";
 import { Table } from "../../api/table";
 import { BaseFunction } from "./__base";
 import { ensureBoolean, ensureNumber, stripTable } from "./__utils";
